RAG-python/requirements.txt
sentence-transformers>=2.3.0
chromadb>=0.4.15
google-generativeai>=0.3.2
numpy>=1.24.3
python-dotenv>=1.0.0
huggingface_hub>=0.20.0 